+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added shaken link to official module link list.
 ### Changed
 - Merged both Unshake functions into a single one. This should make it a lot easier to maintain it in the future and (hopefully) not affect the user in any way.
+- Mark dead will now apply defeated to NPCs and incapacitated to PCs.
 ### Fixed
 - Tester macro failing to create a support effect if a combat is active.
 ### Known issues